+++ app/src/main/kotlin/com/wire/android/ui/WireApp.kt
@@ -1,8 +1,7 @@
 package com.wire.android.ui
 
 import androidx.compose.runtime.Composable
-import com.wire.android.ui.common.UnreadMentionBadge
-import com.wire.android.ui.conversation.Conversation
+import com.wire.android.ui.conversation.all.Conversation
 
 // Here could be an entry point for navigation part, for now I just redirect to ConversationScreen
 @Composable
