+++ app/src/main/kotlin/com/wire/android/ui/conversation/all/ConversationViewModel.kt
@@ -1,6 +1,8 @@
-package com.wire.android.ui.conversation
+package com.wire.android.ui.conversation.all
 
 import androidx.lifecycle.ViewModel
+import com.wire.android.ui.conversation.conversationMockData
+import com.wire.android.ui.conversation.newActivitiesMockData
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
